A little more utils restructuring, update docs for new structure
