--- docs/utility_functions.rst
@@ -1,5 +1,0 @@
-Utility Functions
-=================
-
-.. automodule:: pyuvdata.utils
-  :members:
