+++ src/pyuvdata/uvbeam/beamfits.py
@@ -12,7 +12,7 @@ from docstring_parser import DocstringStyle
 
 from .. import utils
 from ..docstrings import copy_replace_short_description
-from ..utils.file_io import fits as fits_utils
+from ..utils.io import fits as fits_utils
 from . import UVBeam
 
 __all__ = ["BeamFITS"]
