+++ src/pyuvdata/uvcal/calfits.py
@@ -20,7 +20,7 @@ except ImportError:
 
 from .. import utils
 from ..docstrings import copy_replace_short_description
-from ..utils.file_io import fits as fits_utils
+from ..utils.io import fits as fits_utils
 from . import UVCal
 
 __all__ = ["CALFITS"]
