+++ src/pyuvdata/uvcal/ms_cal.py
@@ -13,7 +13,7 @@ from docstring_parser import DocstringStyle
 
 from .. import utils
 from ..docstrings import copy_replace_short_description
-from ..utils.file_io import ms as ms_utils
+from ..utils.io import ms as ms_utils
 from . import UVCal
 
 __all__ = ["MSCal"]
