+++ tests/utils/io/__init__.py
@@ -1,4 +1,4 @@
 # -*- mode: python; coding: utf-8 -*-
 # Copyright (c) 2024 Radio Astronomy Software Group
 # Licensed under the 2-clause BSD License
-"""Tests for utility file_io functions."""
+"""Tests for utility file io functions."""
