+++ tests/uvcal/test_uvcal.py
@@ -15,7 +15,7 @@ import pytest
 from astropy.io import fits
 from astropy.table import Table
 
-import pyuvdata.utils.file_io.fits as fits_utils
+import pyuvdata.utils.io.fits as fits_utils
 from pyuvdata import UVCal, utils
 from pyuvdata.data import DATA_PATH
 from pyuvdata.testing import check_warnings
