+++ tests/uvdata/test_uvfits.py
@@ -12,7 +12,7 @@ import numpy as np
 import pytest
 from astropy.io import fits
 
-import pyuvdata.utils.file_io.fits as fits_utils
+import pyuvdata.utils.io.fits as fits_utils
 from pyuvdata import UVData, utils
 from pyuvdata.data import DATA_PATH
 from pyuvdata.testing import check_warnings
